Creates the PressableModuleBase for DS modules
