+++ src/components/modules/index.tsx
@@ -1,2 +1,3 @@
 export * from "./ModuleIDP";
 export * from "./ModulePaymentNotice";
+export * from "./PressableModuleBase";
